(PC-17806) refactor(auth): move all forgotten password screens in pages folder
