--- src/features/auth/forgottenPassword/ForgottenPassword/index.ts
@@ -1,1 +1,0 @@
-export * from './ForgottenPassword'
