--- src/features/auth/forgottenPassword/ReinitializePassword/index.ts
@@ -1,1 +1,0 @@
-export * from './ReinitializePassword'
