--- src/features/auth/forgottenPassword/ResetPasswordEmailSent/index.ts
@@ -1,1 +1,0 @@
-export * from './ResetPasswordEmailSent'
