--- src/features/auth/forgottenPassword/ResetPasswordExpiredLink/index.ts
@@ -1,1 +1,0 @@
-export * from './ResetPasswordExpiredLink'
